+++ spec/integrations/rubocop_spec.rb
@@ -8,15 +8,4 @@ describe "RuboCop cops" do
       expect(output).to include("1 offense detected")
     end
   end
-
-  context "AggregateFailures" do
-    specify do
-      raise "Remove deprecated AggregateFailures" if TestProf::VERSION >= "1.0"
-
-      output = run_rubocop("aggregate_failures", cop: "RSpec/AggregateFailures")
-
-      expect(output).to include("cop has been renamed")
-      expect(output).to include("1 offense detected")
-    end
-  end
 end